# docs/openapi-public.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Community Feed API",
    "description": "Community-governed Bluesky feed generator where subscribers democratically vote on ranking algorithm weights. Built on AT Protocol.\n\n## Authentication\n- **Governance endpoints** require a session cookie or bearer token from `POST /api/governance/auth/login`.\n- **Admin endpoints** additionally require the caller's DID to be in the `BOT_ADMIN_DIDS` allowlist.\n- **Feed endpoints** are public (called by the Bluesky app). Auth is optional for subscriber tracking.\n- **Transparency endpoints** are public and unauthenticated.",
    "version": "1.1.0",
    "contact": {
      "name": "Community Feed",
      "url": "https://github.com/andrewnordstrom-eng/bluesky-community-feed"
    },
    "license": {
      "name": "MIT"
    }
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "governance_session",
        "description": "Session cookie from POST /api/governance/auth/login"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token from POST /api/governance/auth/login (for CLI/MCP)"
      }
    },
...
  },
  "paths": {
...
    "/api/governance/weights": {
      "get": {
        "summary": "Get current weights",
        "tags": ["Governance"],
        "description": "Returns the active epoch's governance weights, vote count, and metadata.",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "epoch_id": {
                      "type": "integer",
                      "description": "Active epoch ID"
                    },
                    "status": {
                      "type": "string",
                      "description": "Epoch status"
                    },
                    "weights": {
                      "type": "object",
                      "properties": {
                        "recency": {
                          "type": "number"
                        },
                        "engagement": {
                          "type": "number"
                        },
                        "bridging": {
                          "type": "number"
                        },
                        "sourceDiversity": {
                          "type": "number"
                        },
                        "relevance": {
                          "type": "number"
                        }
                      }
                    },
                    "vote_count": {
                      "type": "integer",
                      "description": "Number of votes in this epoch"
                    },
                    "created_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "description": {
                      "type": "string",
                      "nullable": true
                    }
                  },
                  "required": ["epoch_id", "status", "weights", "vote_count"]
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error code (e.g. \"ValidationError\", \"NotFound\")"
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "correlationId": {
                      "type": "string",
                      "description": "Request correlation ID for debugging"
                    }
                  },
                  "required": ["error", "message"]
                }
              }
            }
          }
        }
      }
    },
    "/api/governance/weights/history": {
      "get": {
        "summary": "Weight history",
        "tags": ["Governance"],
        "description": "Returns all epochs with their weights, ordered newest first. Useful for timeline visualization of governance changes.",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "in": "query",
            "name": "limit",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "epochs": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "epoch_id": {
                            "type": "integer"
                          },
                          "status": {
                            "type": "string"
                          },
                          "weights": {
                            "type": "object",
                            "properties": {
                              "recency": {
                                "type": "number"
                              },
                              "engagement": {
                                "type": "number"
                              },
                              "bridging": {
                                "type": "number"
                              },
                              "sourceDiversity": {
                                "type": "number"
                              },
                              "relevance": {
                                "type": "number"
                              }
                            }
                          },
                          "vote_count": {
                            "type": "integer"
                          },
                          "created_at": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "closed_at": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true
                          },
                          "description": {
                            "type": "string",
                            "nullable": true
                          }
                        }
                      }
                    },
                    "total": {
                      "type": "integer",
                      "description": "Number of epochs returned"
                    }
                  },
                  "required": ["epochs", "total"]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error code (e.g. \"ValidationError\", \"NotFound\")"
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "correlationId": {
                      "type": "string",
                      "description": "Request correlation ID for debugging"
                    }
                  },
                  "required": ["error", "message"]
                }
              }
            }
          }
        }
      }
    },
    "/api/governance/weights/compare": {
      "get": {
        "summary": "Compare epoch weights",
        "tags": ["Governance"],
        "description": "Compare weights between two epochs, showing the difference for each component.",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0
            },
            "in": "query",
            "name": "epoch1",
            "required": true
          },
          {
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0
            },
            "in": "query",
            "name": "epoch2",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "epoch1": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "weights": {
                          "type": "object",
                          "properties": {
                            "recency": {
                              "type": "number"
                            },
                            "engagement": {
                              "type": "number"
                            },
                            "bridging": {
                              "type": "number"
                            },
                            "sourceDiversity": {
                              "type": "number"
                            },
                            "relevance": {
                              "type": "number"
                            }
                          }
                        },
                        "status": {
                          "type": "string"
                        }
                      }
                    },
                    "epoch2": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "weights": {
                          "type": "object",
                          "properties": {
                            "recency": {
                              "type": "number"
                            },
                            "engagement": {
                              "type": "number"
                            },
                            "bridging": {
                              "type": "number"
                            },
                            "sourceDiversity": {
                              "type": "number"
                            },
                            "relevance": {
                              "type": "number"
                            }
                          }
                        },
                        "status": {
                          "type": "string"
                        }
                      }
                    },
                    "difference": {
                      "type": "object",
                      "description": "Weight differences (epoch2 − epoch1)",
                      "properties": {
                        "recency": {
                          "type": "number"
                        },
                        "engagement": {
                          "type": "number"
                        },
                        "bridging": {
                          "type": "number"
                        },
                        "sourceDiversity": {
                          "type": "number"
                        },
                        "relevance": {
                          "type": "number"
                        }
                      }
                    }
                  },
                  "required": ["epoch1", "epoch2", "difference"]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error code (e.g. \"ValidationError\", \"NotFound\")"
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "correlationId": {
                      "type": "string",
                      "description": "Request correlation ID for debugging"
                    }
                  },
                  "required": ["error", "message"]
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error code (e.g. \"ValidationError\", \"NotFound\")"
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "correlationId": {
                      "type": "string",
                      "description": "Request correlation ID for debugging"
                    }
                  },
                  "required": ["error", "message"]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "https://feed.yourdomain.com",
      "description": "Production"
    }
  ],
...
}